pyproject.toml
[project]
name = "projet10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "easyocr>=1.7.2",
    "faiss-cpu==1.10.0",
    "langchain==0.3.23",
    "langchain-mistralai>=0.2.12",
    "mistralai>=1.0",
    "openpyxl==3.1.5",
    "pandas==2.2.3",
    "pillow>=9.0.0",
    "pydantic-ai-slim[mistral]>=1.52.0",
    "pymupdf>=1.22.0",
    "pypdf2==3.0.1",
    "python-docx==1.1.2",
    "python-dotenv==1.1.0",
    "ragas>=0.4.3",
    "requests==2.32.3",
    "streamlit==1.44.1",
    "streamlit-feedback==0.1.4",
    "tqdm>=4.67.2",
]
